Move resolveRelations(...) out of the Analyzer
